requirements.txt
streamlit==1.29.0
openai==1.6.1
python-dotenv==1.0.0
speechrecognition==3.10.1
gtts==2.5.0
pydub==0.25.1
langdetect==1.0.9
deep-translator==1.11.4
sounddevice==0.4.6
soundfile==0.12.1
numpy==1.24.3
elevenlabs==0.2.27
